+++ src/core/src/text/index.ts
@@ -1,3 +1,1 @@
-export * from "./line-breaks.js";
-export * from "./source-text-validation.js";
-export * from "./string-comment-scan.js";
+export * from "./source-text.js";
